Add discovered tests groups in reverse order.
CppUTest lists tests (and test groups) in reverse order as they are defined in source code files. Tests were already being added in reverse order to display them in VS in the same order that they are declared in source files, but not test groups, so these were being displayed in VS in reverse order.